Change y position to float

+++ models/models.go
@@ -326,7 +326,7 @@ type (
 		FrontMaterial   bson.ObjectId `bson:"front_material_id" json:"front_material_id"`
 		TempleMaterial  bson.ObjectId `bson:"temple_material_id" json:"temple_material_id"`
 		Scale           float64       `bson:"scale" json:"scale"`
-		YPosition       int16         `bson:"y_position" json:"y_position"`
+		YPosition       float64         `bson:"y_position" json:"y_position"`
 		LeftTempleText  string        `bson:"left_temple_text" json:"left_temple_text"`
 		RightTempleText string        `bson:"right_temple_text" json:"right_temple_text"`
 	}
